WrapperGameServerSpawnLivingEntity created, but incomplete, We need to make our own EntityType enum. (and map all ids per version);
Small fixing in PacketTypes, cleaned up some duplicate types.

+++ src/main/java/io/github/retrooper/packetevents/wrapper/game/client/WrapperGameClientInteractEntity.java
@@ -104,7 +104,7 @@ public class WrapperGameClientInteractEntity extends PacketWrapper<WrapperGameCl
     public void writeData() {
         if (clientVersion.isOlderThan(ClientVersion.v_1_8)) {
             writeInt(entityID);
-            writeByte((byte) type.ordinal());
+            writeByte(type.ordinal());
         } else {
             writeVarInt(entityID);
             writeVarInt(type.ordinal());
