+++ src/main/java/io/github/retrooper/packetevents/wrapper/game/server/WrapperGameServerChatMessage.java
@@ -105,7 +105,7 @@ public class WrapperGameServerChatMessage extends PacketWrapper<WrapperGameServe
 
         //Is the server 1.8+ or is the client 1.8+? (1.7.10 servers support 1.8 clients, and send the chat position for 1.8 clients)
         if (getServerVersion().isNewerThanOrEquals(ServerVersion.v_1_8) || getClientVersion().isNewerThanOrEquals(ClientVersion.v_1_8)) {
-            writeByte((byte) position.ordinal());
+            writeByte(position.ordinal());
         }
 
         if (getServerVersion().isNewerThanOrEquals(ServerVersion.v_1_16)) {
